Fixed the "known caveat" where you had to have an
exact stack of GOLD_INGOTS in your inventory for the Gold Ingot system
to allow you to buy tickets. They know spend just like regular money,
and removes bars as you purchase. If out of money, tells you
"Insufficient funds" and will deplete the remaining balance of the cost
of a ticket from a new stack.